Refacto command packages name
