+++ src/main/java/org/leralix/tan/commands/admin/ReloadCommand.java
@@ -1,4 +1,4 @@
-package org.leralix.tan.commands.adminsubcommand;
+package org.leralix.tan.commands.admin;
 
 import org.bukkit.command.CommandSender;
 import org.bukkit.plugin.Plugin;
