+++ src/main/java/org/leralix/tan/commands/debug/SkipDay.java
@@ -1,24 +1,23 @@
-package org.leralix.tan.commands.debugsubcommand;
+package org.leralix.tan.commands.debug;
 
 import org.bukkit.command.CommandSender;
 import org.leralix.lib.commands.SubCommand;
-import org.leralix.tan.utils.ArchiveUtil;
-import org.leralix.tan.utils.TanChatUtils;
+import org.leralix.tan.tasks.DailyTasks;
 import org.leralix.tan.lang.Lang;
 
-import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
-public class CreateBackup extends SubCommand {
+public class SkipDay extends SubCommand {
 
     @Override
     public String getName() {
-        return "createBackup";
+        return "skipday";
     }
 
     @Override
     public String getDescription() {
-        return Lang.DEBUG_CREATE_BACKUP.get();
+        return Lang.DEBUG_SKIP_DAY.get();
     }
 
     @Override
@@ -28,15 +27,14 @@ public class CreateBackup extends SubCommand {
 
     @Override
     public String getSyntax() {
-        return "/tandebug createBackup";
+        return "/tandebug skipday";
     }
     public List<String> getTabCompleteSuggestions(CommandSender commandSender, String lowerCase, String[] args){
-        return new ArrayList<>();
+        return Collections.emptyList();
     }
     @Override
     public void perform(CommandSender commandSender, String[] args) {
-        ArchiveUtil.archiveFiles();
-        commandSender.sendMessage(TanChatUtils.getTANString() + Lang.COMMAND_GENERIC_SUCCESS.get());
+        DailyTasks.executeMidnightTasks();
     }
-
 }
+
