+++ src/main/java/org/leralix/tan/commands/player/ClaimCommand.java
@@ -1,4 +1,4 @@
-package org.leralix.tan.commands.playersubcommand;
+package org.leralix.tan.commands.player;
 
 import org.bukkit.Chunk;
 import org.bukkit.entity.Player;
