+++ src/main/java/org/leralix/tan/commands/player/OpenGuiCommand.java
@@ -1,4 +1,4 @@
-package org.leralix.tan.commands.playersubcommand;
+package org.leralix.tan.commands.player;
 
 
 import org.leralix.lib.commands.PlayerSubCommand;
