+++ src/main/java/org/leralix/tan/commands/player/SetTownSpawnCommand.java
@@ -1,8 +1,11 @@
-package org.leralix.tan.commands.playersubcommand;
+package org.leralix.tan.commands.player;
 
 import org.bukkit.entity.Player;
 import org.leralix.lib.commands.PlayerSubCommand;
-import org.leralix.tan.storage.TeleportationRegister;
+import org.leralix.tan.dataclass.chunk.ClaimedChunk2;
+import org.leralix.tan.dataclass.chunk.TownClaimedChunk;
+import org.leralix.tan.enums.RolePermission;
+import org.leralix.tan.storage.stored.NewClaimedChunkStorage;
 import org.leralix.tan.storage.stored.PlayerDataStorage;
 import org.leralix.tan.storage.stored.TownDataStorage;
 import org.leralix.tan.utils.TanChatUtils;
@@ -13,22 +16,22 @@ import org.leralix.tan.lang.Lang;
 import java.util.Collections;
 import java.util.List;
 
-public class TownSpawnCommand extends PlayerSubCommand {
+public class SetTownSpawnCommand extends PlayerSubCommand {
     @Override
     public String getName() {
-        return "spawn";
+        return "setspawn";
     }
 
     @Override
     public String getDescription() {
-        return Lang.SPAWN_COMMAND_DESC.get();
+        return Lang.SET_SPAWN_COMMAND_DESC.get();
     }
-    public int getArguments(){ return 1;}
 
+    public int getArguments(){ return 1;}
 
     @Override
     public String getSyntax() {
-        return "/tan spawn";
+        return "/tan setspawn";
     }
     @Override
     public List<String> getTabCompleteSuggestions(Player player, String lowerCase, String[] args){
@@ -50,22 +53,29 @@ public class TownSpawnCommand extends PlayerSubCommand {
             return;
         }
 
+        //No permission
         TownData townData = TownDataStorage.getInstance().get(player);
+
+        if(!townData.doesPlayerHavePermission(playerStat, RolePermission.TOWN_ADMINISTRATOR)){
+            player.sendMessage(TanChatUtils.getTANString() + Lang.PLAYER_NO_PERMISSION.get());
+            return;
+        }
+
+
         //Spawn Unlocked
         if(townData.isSpawnLocked()){
             player.sendMessage(TanChatUtils.getTANString() + Lang.SPAWN_NOT_UNLOCKED.get());
             return;
         }
 
-        //Spawn set
-        if(!townData.isSpawnSet()){
-            player.sendMessage(TanChatUtils.getTANString() + Lang.SPAWN_NOT_SET.get());
+        ClaimedChunk2 currentChunk = NewClaimedChunkStorage.getInstance().get(player.getLocation().getChunk());
+        if(!(currentChunk instanceof TownClaimedChunk townChunk && townChunk.getTown().equals(townData))){
+            player.sendMessage(TanChatUtils.getTANString() + Lang.SPAWN_NEED_TO_BE_IN_CHUNK.get());
             return;
         }
 
-        TeleportationRegister.teleportToTownSpawn(playerStat, townData);
-
-
+        townData.setSpawn(player.getLocation());
+        player.sendMessage(TanChatUtils.getTANString() + Lang.SPAWN_SET_SUCCESS.get());
     }
 
 }
